+++ extension.js
@@ -17,8 +17,12 @@ function activate(context) {
 	context.subscriptions.push(disposable);
 }
 
-// this method is called when your extension is deactivated
-function deactivate() { }
+function deactivate() {
+	const statusBarItem = Quotes.getStatusBarItem()
+	if (statusBarItem) {
+		statusBarItem.dispose()
+	}
+}
 
 module.exports = {
 	activate,